refactor: 보안 관련 설정파일 디렉터리 구조 변경
